feat: implement make_encoder; add common.py
